requirements.txt
notion-client>=2.2.0
python-frontmatter>=1.1.0
ruamel.yaml>=0.18.0
markdown-it-py>=3.0.0
mdit-py-plugins>=0.4.0
jsonschema>=4.21.0
